Add central storage manager
This does not yet make crafter manager use this central place.  Coming soon.

This does add a new file; it is added to the installer.
