+++ BIM/Functions/VariableStorage.lua
@@ -17,13 +17,22 @@ local function getEnv(key)
     end
 end
 
+--- Call when adding a new item or on first construction
+local function saveTable(table, where)
+    local file = fs.open(where, "w")
+    assert(file, "Unable to open " .. where .. " for writing.")
+
+    file.write("return " .. textutils.serialise(table))
+    file:close()
+end
+
 local detailsLocation = "/" .. projectName .. "/" .. projectName .. "_itemDetailsMap.lua"
 local reqLocation = "/" .. projectName .. "." .. projectName .. "_itemDetailsMap"
-local dMap = {}
+local itemDetailsMap = {}
 if fs.exists(detailsLocation) then -- Protection against malformed map file
     local ok, result = pcall(require, reqLocation)
     if ok and type(result) == "table" then
-        dMap = result
+        itemDetailsMap = result
     end
 end
 
@@ -53,12 +62,38 @@ do
     env['Name'] = projectName
 end
 
-return {
+local function saveItemDetails()
+    saveTable(itemDetailsMap, detailsLocation)
+end
+
+local function setItemDetail(name, chest, slot, suppressSave)
+    if not itemDetailsMap[name] then
+        local detail = chest.getItemDetail(slot)
+        if detail and detail.displayName then
+            -- details.itemGroups is apparently deprecated and the wiki says it's no longer available. but I see it in mc 1.21.7
+            itemDetailsMap[name] = {
+                tags = detail.tags,
+                maxCount = detail.maxCount,
+                displayName = detail.displayName
+            }
+            if not suppressSave then
+                saveItemDetails()
+            end
+            return true
+        end
+    end
+    return false
+end
+
+local Vs = {
     chests = chests, -- key of an itemname, value example {['side']=peripheral.getName(chest),['slot']=j,['count']=item.count,['name']=name}
     list = itemlist, -- list of items in system { {count:int, displayName:string, id:string} }
     name = projectName,
-    itemDetailsMapLocation = detailsLocation,
-    itemDetailsMap = dMap,
+    setItemDetail = setItemDetail,
+    itemDetailsMap = itemDetailsMap,
+    saveItemDetails = saveItemDetails,
     getEnv = getEnv,
     setKeyEnv = setKeyEnv
 }
+
+return Vs
