+++ BIM/InventoryManager.lua
@@ -1,4 +1,9 @@
 --todo consider making clicked item bg flash instead of stay selected indefinitely
+--todo test if thousands of items get rounded to the thousand
+--todo consider allow clicking in the middle of sortString to add text in the center.
+--todo add scrolling when dragging scroll bar
+--todo work with itemGroups
+--todo figure out if selected needs to be a table
 
 --#region Locals--
 local settingPath = Vs.name .. '/' .. Vs.name .. '.settings'
@@ -16,13 +21,12 @@ local sortDisplay = {
     "Amount " .. string.char(0x1F)
 }
 
-local filtered = {} -- Same structure as Vs.list
+local filtered = {} -- Same structure as Storage.list, except filtered with search bar
 local clickList = {}
 local sClickList = {}
 local scrollIndex = 0
 local selected = {}
 local buffer
-local chests = {} --- A list of all chest peripherals in the system
 local monitor = nil
 
 local mainScreen = term.current()
@@ -42,35 +46,7 @@ local searchText = ""
 local colAmount
 --#endregion Locals--
 
---- Call when adding a new item or on first construction
-local function saveTable(table, where)
-    local file = fs.open(where, "w")
-    assert(file, "Unable to open ".. where .. " for writing.")
-
-    file.write("return " .. textutils.serialise(table))
-    file:close()
-end
-
 local function trim(s) return s:match("^%s*(.-)%s*$") end
-
-local function setItemDetail(name, chest, slot, suppressSave)
-    if not Vs.itemDetailsMap[name] then
-        local detail = chest.getItemDetail(slot)
-        if detail and detail.displayName then
-            -- details.itemGroups is apparently deprecated and the wiki says it's no longer available. but I see it in mc 1.21.7
-            Vs.itemDetailsMap[name] = {
-                tags = detail.tags,
-                maxCount = detail.maxCount,
-                displayName = detail.displayName
-            }
-            if not suppressSave then
-                saveTable(Vs.itemDetailsMap, Vs.itemDetailsMapLocation)
-            end
-            return true
-        end
-    end
-    return false
-end
 
 --#region Functions--
 local function filter(fList)
@@ -108,110 +84,8 @@ local function filter(fList)
 end
 
 local function sortList()
-    filter(Vs.list)
+    filter(Storage.list)
     table.sort(filtered, listSort[sortIndex])
-end
-
----Maps out where items are
-local function scanStorage()
-    local chestlist = {}
-    local foundNewItemType = false
-    local itemList = {}
-    for _, chest in ipairs(chests) do
-        local items = chest.list()
-        for slot, item in pairs(items) do
-            local name = item.name -- i.e. minecraft:bone
-            chestlist[name] = chestlist[name] or {}
-
-            if setItemDetail(name, chest, slot, true) then
-                foundNewItemType = true
-            end
-
-            table.insert(chestlist[name],
-                {
-                    side = peripheral.getName(chest),
-                    slot = slot,
-                    count = item.count,
-                    name = name, -- same as id
-                }
-            )
-
-            itemList[name] = itemList[name] or 0
-            itemList[name] = itemList[name] + item.count
-        end
-    end
-
-    local newItemList = {}
-    for name, count in pairs(itemList) do
-        table.insert(newItemList, {name = name, count = count})
-    end
-    Vs.list = newItemList
-    filtered = newItemList
-    sortList()
-
-    Vs.chests = chestlist
-    if foundNewItemType then
-        saveTable(Vs.itemDetailsMap, Vs.itemDetailsMapLocation)
-    end
-
-end
-
-local function sortItems()
-    for _, item in ipairs(Vs.list) do
-        local maxCount = Vs.itemDetailsMap[item.name].maxCount
-        local maxSlotsNeeded = math.ceil(item.count / maxCount)
-        local itemChests = Vs.chests[item.name]
-        if #itemChests > maxSlotsNeeded then
-            -- Make a copy of itemChests indices to avoid issues when removing
-            local frontChest = 1
-            while frontChest <= #itemChests do
-                local frontSlotData = itemChests[frontChest]
-                if frontSlotData.count < maxCount then
-                    local frontChestPeripheral = peripheral.wrap(frontSlotData.side)
-                    assert(frontChestPeripheral, "Peripheral not found.  Please report bug.")
-
-                    -- Go from last chest with item and move it to front
-                    for backChest = #itemChests, frontChest + 1, -1 do
-                        local backSlotData = itemChests[backChest]
-                        local backSide = backSlotData.side
-                        local transferred = frontChestPeripheral.pullItems(
-                            backSide,
-                            backSlotData.slot,
-                            64,
-                            frontSlotData.slot
-                        )
-                        backSlotData.count = backSlotData.count - transferred
-                        frontSlotData.count = frontSlotData.count + transferred
-                        if backSlotData.count <= 0 then
-                            table.remove(itemChests, backChest)
-                        end
-                        if frontSlotData.count == maxCount then
-                            break
-                        end
-                    end
-                end
-                -- Only increment if we didn't remove the current frontChest
-                if frontSlotData.count == maxCount or frontChest == #itemChests then
-                    frontChest = frontChest + 1
-                end
-            end
-        end
-    end
-end
-
-local function pushBufferItemToStorage(chestName, fromSlotIndex, toSlotIndex, itemId)
-    local pushed = buffer.pushItems(chestName, fromSlotIndex, 64, toSlotIndex)
-    if pushed > 0 then
-        -- Update Vs.chests
-        Vs.chests[itemId] = Vs.chests[itemId] or {}
-        table.insert(Vs.chests[itemId], {
-            side = chestName,
-            slot = toSlotIndex,
-            count = pushed,
-            name = itemId
-        })
-    end
-    return pushed
 end
 
 local function storeItems()
@@ -235,54 +109,8 @@ local function storeItems()
                     end
                 end
 
-                -- Get buffer inventory
-                local bufferItems = buffer.list()
-                for slot, item in pairs(bufferItems) do
-                    local itemId = item.name
-                    local itemCount = item.count
-                    setItemDetail(itemId, buffer, slot)
-
-                    local chestSlots = Vs.chests[itemId] or {}
-                    local remaining = itemCount
-                    -- Fill non-full stacks
-                    for _, chestSlot in ipairs(chestSlots) do
-                        if remaining <= 0 then break end
-                        local pushed = buffer.pushItems(chestSlot.side, slot, 64, chestSlot.slot)
-                        chestSlot.count = chestSlot.count + pushed
-                        remaining = remaining - pushed
-                    end
-
-                    -- Put remaining items in empty slots
-                    if remaining > 0 then
-                        for _, chest in ipairs(chests) do
-                            local chestName = peripheral.getName(chest)
-                            local chestInv = chest.list()
-
-                            for chestSlotNum = 1, chest.size() do
-                                if not chestInv[chestSlotNum] then
-                                    local pushed = pushBufferItemToStorage(chestName, slot, chestSlotNum, itemId)
-                                    remaining = remaining - pushed
-                                    if remaining <= 0 then break end
-                                end
-                            end
-                            if remaining <= 0 then break end
-                        end
-                    end
-
-                    -- Update Vs.list
-                    local found = false
-                    for _, entry in ipairs(Vs.list) do
-                        if entry.name == itemId then
-                            entry.count = entry.count + itemCount
-                            found = true
-                            break
-                        end
-                    end
-                    if not found then
-                        table.insert(Vs.list, { count = itemCount, name = itemId })
-                    end
-                end
-                filter(Vs.list)
+                Storage:storeBuffer()
+                filter(Storage.list)
                 sortList()
                 clickList = Um.Print(filtered, selected, scrollIndex, scrollBar, screen, colAmount)
                 if monitor then sClickList = Um.Print(filtered, selected, 0, nil, secondScreen, colAmount) end
@@ -311,7 +139,7 @@ end
 local function loopSort()
     while true do
         if buffer then
-            sortItems()
+            Storage:sortStorage()
 
             scrollIndex = math.min(
                 math.max(scrollIndex, 0),
@@ -337,52 +165,11 @@ end
 ---@param percentOfStack number 0-1 how much of a stack to pull
 local function dropItem(id, percentOfStack)
     if id == nil or filtered[id] == nil then return nil end
-    local itemName = filtered[id].name
-    local chestData = Vs.chests[itemName]
-    if not chestData or #chestData == 0 then return nil end
-
-    -- Cache maxCount
-    local maxCount = 64
-    local chest = peripheral.wrap(chestData[1].side)
-    if chest then
-        setItemDetail(itemName, chest, chestData[1].slot)
-    end
-
-    local amountToPull = math.ceil(maxCount * percentOfStack)
-    local stack = 0
     selected = { filtered[id] }
-
-    for i = #chestData, 1, -1 do
-        local data = chestData[i]
-        if data.count > 0 and stack < amountToPull then
-            local amount = math.min(amountToPull - stack, data.count)
-            local transferred = buffer.pullItems(data.side, data.slot, amount)
-
-            -- Remove from Vs.list to prevent items from re-appearing when sorting
-            -- todo O(n) check if this can be faster; may not need to after refactoring to stop need for sorting
-            -- If I only update when the #Vs.chests[item] == 0 then I won't
-            for l = #Vs.list, 1, -1 do
-                local listItem = Vs.list[l]
-                if listItem.name == itemName then
-                    listItem.count = listItem.count - transferred
-                end
-                if listItem.count == 0 then
-                    table.remove(Vs.list, l)
-                    table.remove(filtered, id) -- instantly remove from viewed list if none left
-                end
-            end
-
-            -- update variable storage chests
-            data.count = data.count - transferred
-            if data.count == 0 then
-                table.remove(chestData, i)
-            end
-
-            stack = stack + transferred
-            if stack >= amountToPull then
-                break
-            end
-        end
+    local itemName = filtered[id].name
+    local _, droppedLastItem = Storage:retrieveItem(itemName, percentOfStack)
+    if droppedLastItem then
+        table.remove(filtered, id) -- instantly remove from viewed list if none left
     end
 
     Um.Print(filtered, selected, scrollIndex, scrollBar, screen, colAmount)
@@ -397,6 +184,10 @@ local function dropItem(id, percentOfStack)
 end
 
 local function loopPrint()
+    local keyscroll = {
+        [keys.getName(keys.up)] = -1,
+        [keys.getName(keys.down)] = 1
+    }
     while true do
         if buffer then
             local event = { os.pullEvent() }
@@ -405,6 +196,19 @@ local function loopPrint()
                     scrollIndex = scrollIndex + event[2]
                     printScreen()
                 end
+            elseif event[1] == "key" then
+                -- local validKey = keyscroll[keys.getName(event[2])]
+                -- if validKey then
+                --     if scrollIndex ~= math.min(math.max(scrollIndex + validKey, 0), math.max(math.ceil(#filtered / colAmount) - screenSize[2], 0)) then
+                --         scrollIndex = scrollIndex + event[2]
+                --         printScreen()
+                --     end
+                -- end
+                -- if event[2] == keys.getName(keys.up) then
+
+                -- elseif event[2] == keys.getName(keys.down) then
+                -- end
+                -- printScreen()
             elseif event[1] == 'mouse_click' and event[4] >= 2 and event[3] < select(1, term.getSize()) then
                 local dropAmount = { 1, 0.5, 0.01 }
                 dropItem(Um.Click(clickList, event[3], event[4]), dropAmount[event[2]])
@@ -421,17 +225,6 @@ end
 
 local function loadEnv()
     buffer = peripheral.wrap(Vs.getEnv('Buffer'))
-    chests = {peripheral.find(Vs.getEnv("Inventories"))}
-    local cardinal = { left = true, right = true, top = true, bottom = true, front = true, back = true }
-
-    for i = #chests, 1, -1 do
-        local chestName = peripheral.getName(chests[i])
-        -- Storage should not include inventories adjacent to turtle or buffer inventory
-        if cardinal[chestName] or chestName == Vs.getEnv("Buffer") then
-            table.remove(chests, i)
-        end
-    end
-
     colAmount = tonumber(Vs.getEnv('Columns'))
 
     monitor = peripheral.wrap(Vs.getEnv('Monitor'))
@@ -455,7 +248,9 @@ local function loadEnv()
     settings.set(Vs.name .. ".sortIndex", sortIndex)
     settings.save(settingPath)
 
-    scanStorage()
+    Storage:scanStorage()
+    filtered = Storage.list
+    sortList()
 end
 
 local function loopEnv()
