+++ Startup/StartBIM.lua
@@ -1,7 +1,9 @@
-ProjectName = 'BIM'
-Programs = { 'Inventory', 'Crafter', 'Settings' }
-local Um = require('/' .. ProjectName .. '/Functions/UiManager')
-local Vs = require('/' .. ProjectName .. '/Functions/VariableStorage')
+ProjectName = "BIM"
+Programs = { "Inventory", "Crafter", "Settings" }
+local Um = require('/' .. ProjectName .. "/Functions/UiManager")
+local SS = require('/' .. ProjectName .. "/Functions/StorageSystem")
+local Vs = require('/' .. ProjectName .. "/Functions/VariableStorage")
+SS:init(Vs)
 Um.setVs(Vs)
 
 for _, value in ipairs(Programs) do
@@ -10,6 +12,7 @@ for _, value in ipairs(Programs) do
             {
                 Um = Um,
                 Vs = Vs,
+                Storage = SS,
                 require = require,
                 multishell = multishell
             },
